rename hooks to proper react convention
